Refactoring & Cleanup
Refactored project for pnpm, broke down larger files into smaller ones and improved quality of code & comments
