+++ frontend/public/index.html
@@ -22,7 +22,7 @@
 
       Unlike "/favicon.ico" or "favicon.ico", "%PUBLIC_URL%/favicon.ico" will
       work correctly both with client-side routing and a non-root public URL.
-      Learn how to configure a non-root public URL by running `npm run build`.
+      Learn how to configure a non-root public URL by running `pnpm run build`.
     -->
     <title>Canvas Assignment Scheduler</title>
   </head>
@@ -36,8 +36,8 @@
       You can add webfonts, meta tags, or analytics to this file.
       The build step will place the bundled scripts into the <body> tag.
 
-      To begin the development, run `npm start` or `yarn start`.
-      To create a production bundle, use `npm run build` or `yarn build`.
+      To begin the development, run `pnpm start`.
+      To create a production bundle, use `pnpm run build`.
     -->
   </body>
 </html>
