+++ frontend/src/constants/index.ts
@@ -1,0 +1,3 @@
+// shared constants for the canvas assignment scheduler app
+
+export const API_BASE = 'http://localhost:8000/api';
